+++ core/common/src/main/kotlin/com/github/jcornaz/miop/Operators.kt
@@ -339,3 +339,39 @@ public fun <E> ReceiveChannel<E>.windowed(size: Int, step: Int = 1, partialWindo
 @ObsoleteCoroutinesApi
 @Suppress("NOTHING_TO_INLINE")
 public inline fun <E> ReceiveChannel<E>.chunked(size: Int): ReceiveChannel<List<E>> = windowed(size, size, true)
+
+/**
+ * Send downstream each intermediate result of accumulating the emitted elements using the given [operation] function, and starting with a [initial].
+ *
+ * Example: `receiveChannelOf(1, 2, 3).san(0) { acc, element -> acc + element }` will emit: `[0, 1, 3, 6]`
+ */
+@ObsoleteCoroutinesApi
+public fun <T, R> ReceiveChannel<T>.scan(initial: R, operation: (acc: R, T) -> R): ReceiveChannel<R> = transform { input, output ->
+    output.send(initial)
+    var result = initial
+    input.consumeEach {
+        result = operation(result, it)
+        output.send(result)
+    }
+}
+
+/**
+ * Send downstream each intermediate result of accumulating the emitted elements using the given [operation] function.
+ *
+ * Example: `receiveChannelOf(1, 2, 3).san { acc, element -> acc + element }` will emit: `[1, 3, 6]`
+ */
+@ObsoleteCoroutinesApi
+public fun <T> ReceiveChannel<T>.scan(operation: (acc: T, T) -> T): ReceiveChannel<T> = transform { input, output ->
+    var result = try {
+        receive()
+    } catch (closedChannel: ClosedReceiveChannelException) {
+        return@transform
+    }
+
+    output.send(result)
+
+    input.consumeEach {
+        result = operation(result, it)
+        output.send(result)
+    }
+}